Update evolution benchmark
