+++ examples/benchmarks/evolution.py
@@ -1,10 +1,9 @@
 """Adiabatic evolution for the Ising Hamiltonian using linear scaling."""
-import os
 import argparse
-import json
 import time
 import qibo
 from qibo import callbacks, hamiltonians, models
+from utils import parse_accelerators, BenchmarkLogger
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--nqubits", default=4, type=int)
@@ -12,41 +11,11 @@ parser.add_argument("--dt", default=1e-2, type=float)
 parser.add_argument("--solver", default="exp", type=str)
 parser.add_argument("--dense", action="store_true")
 parser.add_argument("--accelerators", default=None, type=str)
-parser.add_argument("--backend", default="qibotf", type=str)
+parser.add_argument("--backend", default="qibojit", type=str)
 parser.add_argument("--filename", default=None, type=str)
 
 
-def parse_accelerators(accelerators):
-    """Transforms string that specifies accelerators to dictionary.
-
-    The string that is parsed has the following format:
-        n1device1,n2device2,n3device3,...
-    and is transformed to the dictionary:
-        {'device1': n1, 'device2': n2, 'device3': n3, ...}
-
-    Example:
-        2/GPU:0,2/GPU:1 --> {'/GPU:0': 2, '/GPU:1': 2}
-    """
-    if accelerators is None:
-        return None
-
-    def read_digit(x):
-        i = 0
-        while x[i].isdigit():
-            i += 1
-        return x[i:], int(x[:i])
-
-    acc_dict = {}
-    for entry in accelerators.split(","):
-        device, n = read_digit(entry)
-        if device in acc_dict:
-            acc_dict[device] += n
-        else:
-            acc_dict[device] = n
-    return acc_dict
-
-
-def main(nqubits, dt, solver, backend, trotter=False, accelerators=None,
+def main(nqubits, dt, solver, backend, dense=False, accelerators=None,
          filename=None):
     """Performs adiabatic evolution with critical TFIM as the "hard" Hamiltonian."""
     qibo.set_backend(backend)
@@ -54,19 +23,9 @@ def main(nqubits, dt, solver, backend, trotter=False, accelerators=None,
         dense = False
         solver = "exp"
 
-    if filename is not None:
-        if os.path.isfile(filename):
-            with open(filename, "r") as file:
-                logs = json.load(file)
-            print("Extending existing logs from {}.".format(filename))
-        else:
-            print("Creating new logs in {}.".format(filename))
-            logs = []
-    else:
-        logs = []
-
+    logs = BenchmarkLogger(filename)
     logs.append({
-        "nqubits": nqubits, "dt": dt, "solver": solver, "trotter": trotter,
+        "nqubits": nqubits, "dt": dt, "solver": solver, "dense": dense,
         "backend": qibo.get_backend(), "precision": qibo.get_precision(),
         "device": qibo.get_device(), "threads": qibo.get_threads(),
         "accelerators": accelerators
@@ -76,11 +35,11 @@ def main(nqubits, dt, solver, backend, trotter=False, accelerators=None,
     print("Backend:", logs[-1]["backend"])
 
     start_time = time.time()
-    h0 = hamiltonians.X(nqubits, trotter=trotter)
-    h1 = hamiltonians.TFIM(nqubits, h=1.0, trotter=trotter)
+    h0 = hamiltonians.X(nqubits, dense=dense)
+    h1 = hamiltonians.TFIM(nqubits, h=1.0, dense=dense)
     logs[-1]["hamiltonian_creation_time"] = time.time() - start_time
     print(f"\nnqubits = {nqubits}, solver = {solver}")
-    print(f"trotter = {trotter}, accelerators = {accelerators}")
+    print(f"dense = {dense}, accelerators = {accelerators}")
     print("Hamiltonians created in:", logs[-1]["hamiltonian_creation_time"])
 
     start_time = time.time()
@@ -93,10 +52,7 @@ def main(nqubits, dt, solver, backend, trotter=False, accelerators=None,
     final_psi = evolution(final_time=1.0)
     logs[-1]["simulation_time"] = time.time() - start_time
     print("Simulation time:", logs[-1]["simulation_time"])
-
-    if filename is not None:
-        with open(filename, "w") as file:
-            json.dump(logs, file)
+    logs.dump()
 
 
 if __name__ == "__main__":
