+++ examples/benchmarks/vqe.py
@@ -1,13 +1,12 @@
 """
-Testing Variational Quantum Eigensolver.
+Benchmark Variational Quantum Eigensolver.
 """
-import os
 import argparse
 import time
 import numpy as np
 import qibo
 from qibo import gates, models, hamiltonians, K
-from logger import BenchmarkLogger
+from utils import BenchmarkLogger
 
 
 parser = argparse.ArgumentParser()
